Remove grpc-status-details when no pb exists (#11249)
* remove pb status

* fix

* fix

* fix

* fix
