+++ dubbo-rpc/dubbo-rpc-triple/src/main/java/org/apache/dubbo/rpc/protocol/tri/stream/AbstractStream.java
@@ -18,6 +18,7 @@
 package org.apache.dubbo.rpc.protocol.tri.stream;
 
 import org.apache.dubbo.common.threadpool.serial.SerializingExecutor;
+import org.apache.dubbo.common.utils.ClassUtils;
 import org.apache.dubbo.rpc.model.FrameworkModel;
 
 import java.util.concurrent.Executor;
@@ -30,8 +31,26 @@ public abstract class AbstractStream implements Stream {
     protected final Executor executor;
     protected final FrameworkModel frameworkModel;
 
+
+    private static final boolean HAS_PROTOBUF = hasProtobuf();
+
     public AbstractStream(Executor executor, FrameworkModel frameworkModel) {
         this.executor = new SerializingExecutor(executor);
         this.frameworkModel = frameworkModel;
+    }
+
+
+    public static boolean getGrpcStatusDetailEnabled() {
+        return HAS_PROTOBUF;
+    }
+
+
+    private static boolean hasProtobuf() {
+        try {
+            ClassUtils.forName("com.google.protobuf.Message");
+            return true;
+        } catch (ClassNotFoundException ignore) {
+            return false;
+        }
     }
 }
