+++ dubbo-rpc/dubbo-rpc-triple/src/main/java/org/apache/dubbo/rpc/protocol/tri/stream/TripleClientStream.java
@@ -21,6 +21,8 @@ import org.apache.dubbo.common.logger.ErrorTypeAwareLogger;
 import org.apache.dubbo.common.logger.LoggerFactory;
 import org.apache.dubbo.rpc.TriRpcStatus;
 import org.apache.dubbo.rpc.model.FrameworkModel;
+import org.apache.dubbo.rpc.protocol.tri.ClassLoadUtil;
+import org.apache.dubbo.rpc.protocol.tri.ExceptionUtils;
 import org.apache.dubbo.rpc.protocol.tri.TripleHeaderEnum;
 import org.apache.dubbo.rpc.protocol.tri.command.CancelQueueCommand;
 import org.apache.dubbo.rpc.protocol.tri.command.DataQueueCommand;
@@ -36,6 +38,10 @@ import org.apache.dubbo.rpc.protocol.tri.transport.TripleCommandOutBoundHandler;
 import org.apache.dubbo.rpc.protocol.tri.transport.TripleHttp2ClientResponseHandler;
 import org.apache.dubbo.rpc.protocol.tri.transport.WriteQueue;
 
+import com.google.protobuf.Any;
+import com.google.rpc.DebugInfo;
+import com.google.rpc.ErrorInfo;
+import com.google.rpc.Status;
 import io.netty.buffer.ByteBuf;
 import io.netty.channel.Channel;
 import io.netty.channel.ChannelFuture;
@@ -46,10 +52,15 @@ import io.netty.handler.codec.http2.Http2StreamChannelBootstrap;
 import io.netty.util.ReferenceCountUtil;
 import io.netty.util.concurrent.Future;
 
+import java.io.IOException;
 import java.net.SocketAddress;
 import java.nio.charset.StandardCharsets;
+import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 import java.util.concurrent.Executor;
+
+import static org.apache.dubbo.common.constants.LoggerCodeConstants.PROTOCOL_FAILED_RESPONSE;
 
 
 /**
@@ -195,7 +206,14 @@ public class TripleClientStream extends AbstractStream implements ClientStream {
             final Map<String, Object> attachments = headersToMap(trailers, () -> {
                 return reserved.get(TripleHeaderEnum.TRI_HEADER_CONVERT.getHeader());
             });
-            listener.onComplete(status, attachments, reserved);
+            final TriRpcStatus detailStatus;
+            final TriRpcStatus statusFromTrailers = getStatusFromTrailers(reserved);
+            if (statusFromTrailers != null) {
+                detailStatus = statusFromTrailers;
+            } else {
+                detailStatus = status;
+            }
+            listener.onComplete(detailStatus, attachments, reserved);
         }
 
         private TriRpcStatus validateHeaderStatus(Http2Headers headers) {
@@ -309,6 +327,64 @@ public class TripleClientStream extends AbstractStream implements ClientStream {
             }
             return status.appendDescription(
                 "missing GRPC status, inferred error from HTTP status code");
+        }
+
+
+        private TriRpcStatus getStatusFromTrailers(Map<String, String> metadata) {
+            if (null == metadata) {
+                return null;
+            }
+            if (!getGrpcStatusDetailEnabled()){
+                return null;
+            }
+            // second get status detail
+            if (!metadata.containsKey(TripleHeaderEnum.STATUS_DETAIL_KEY.getHeader())) {
+                return null;
+            }
+            final String raw = (metadata.remove(TripleHeaderEnum.STATUS_DETAIL_KEY.getHeader()));
+            byte[] statusDetailBin = StreamUtils.decodeASCIIByte(raw);
+            ClassLoader tccl = Thread.currentThread().getContextClassLoader();
+            try {
+                final Status statusDetail = Status.parseFrom(statusDetailBin);
+                List<Any> detailList = statusDetail.getDetailsList();
+                Map<Class<?>, Object> classObjectMap = tranFromStatusDetails(detailList);
+
+                // get common exception from DebugInfo
+                TriRpcStatus status = TriRpcStatus.fromCode(statusDetail.getCode())
+                    .withDescription(TriRpcStatus.decodeMessage(statusDetail.getMessage()));
+                DebugInfo debugInfo = (DebugInfo) classObjectMap.get(DebugInfo.class);
+                if (debugInfo != null) {
+                    String msg = ExceptionUtils.getStackFrameString(
+                        debugInfo.getStackEntriesList());
+                    status = status.appendDescription(msg);
+                }
+                return status;
+            } catch (IOException ioException) {
+                return null;
+            } finally {
+                ClassLoadUtil.switchContextLoader(tccl);
+            }
+
+        }
+
+
+        private Map<Class<?>, Object> tranFromStatusDetails(List<Any> detailList) {
+            Map<Class<?>, Object> map = new HashMap<>(detailList.size());
+            try {
+                for (Any any : detailList) {
+                    if (any.is(ErrorInfo.class)) {
+                        ErrorInfo errorInfo = any.unpack(ErrorInfo.class);
+                        map.putIfAbsent(ErrorInfo.class, errorInfo);
+                    } else if (any.is(DebugInfo.class)) {
+                        DebugInfo debugInfo = any.unpack(DebugInfo.class);
+                        map.putIfAbsent(DebugInfo.class, debugInfo);
+                    }
+                    // support others type but now only support this
+                }
+            } catch (Throwable t) {
+                LOGGER.error(PROTOCOL_FAILED_RESPONSE, "", "", "tran from grpc-status-details error", t);
+            }
+            return map;
         }
 
         @Override
