+++ dubbo-rpc/dubbo-rpc-triple/src/main/java/org/apache/dubbo/rpc/protocol/tri/stream/TripleServerStream.java
@@ -182,6 +182,9 @@ public class TripleServerStream extends AbstractStream implements ServerStream {
         String grpcMessage = getGrpcMessage(rpcStatus);
         grpcMessage = TriRpcStatus.encodeMessage(TriRpcStatus.limitSizeTo1KB(grpcMessage));
         headers.set(TripleHeaderEnum.MESSAGE_KEY.getHeader(), grpcMessage);
+        if (!getGrpcStatusDetailEnabled()) {
+            return headers;
+        }
         Status.Builder builder = Status.newBuilder().setCode(rpcStatus.code.code)
             .setMessage(grpcMessage);
         Throwable throwable = rpcStatus.cause;
